Eliminate asic_task and ASIC_jobs_queue
Now there is only one queue, the stratum_queue. When a new stratum message is put in the stratum_queue, it's taken out and in case of a `mining.notify`, sent directly to the ASIC. If there's no mining.notify, the dequeue times out based on asic_job_frequency, extranonce_2 is incremented and the jobs is sent to the ASIC. There is no difference between abandon_work or not.

This eliminated rejected shares completely in testing, the only rejections left are stale jobs.